+++ src/glegacy/object_collide.go
@@ -87,5 +87,5 @@ func wrapObjectCollideParseC(ptr unsafe.Pointer) server.ObjectParseFunc {
 func Nox_call_objectType_parseCollide_go(a1 unsafe.Pointer, a2 string, a3 unsafe.Pointer) int {
 	cstr := CString(a2)
 	defer StrFree(cstr)
-	return int(asFuncT[func(*byte, unsafe.Pointer) int32](uintptr(a1))(cstr, a3))
+	return int(asFuncT[func(*byte, unsafe.Pointer) int32](a1)(cstr, a3))
 }
